Moved cppcheck suppressions to inline in the code instead of in a file (issue #11).

+++ source/TrickHLA/Attribute.cpp
@@ -1691,7 +1691,7 @@ void Attribute::decode_opaque_data_from_buffer() // RETURN: -- None.
       }
 
       // Do a sanity check on the decoded length, it should not be negative.
-      if ( decoded_length < 0 ) {
+      if ( decoded_length < 0 ) {      // cppcheck-suppress [knownConditionTrueFalse]
          send_hs( stderr, "Attribute::decode_opaque_data_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d < 0, will use 0 instead.%c",
                   __LINE__, FOM_name, decoded_length, THLA_NEWLINE );
@@ -1706,7 +1706,7 @@ WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d < 0, wi
       } else {
          data_buff_size = 0;
       }
-      if ( (size_t)decoded_length > data_buff_size ) {
+      if ( (size_t)decoded_length > data_buff_size ) {   // cppcheck-suppress [unsignedLessThanZero]
          send_hs( stderr, "Attribute::decode_opaque_data_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d > data buffer \
 size %d, will use the data buffer size instead.%c",
@@ -2340,7 +2340,7 @@ void Attribute::decode_string_from_buffer() // RETURN: -- None.
 
          // Do a sanity check on the decoded length, it should not be negative.
          size_t length;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_UNICODE_STRING attribute '%s' (trick_name: '%s'), decoded length %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, ( ( trick_name != NULL ) ? trick_name : "NULL" ),
@@ -2365,7 +2365,7 @@ WARNING: For ENCODING_TYPE_UNICODE_STRING attribute '%s' (trick_name: '%s'), dec
                // data_buff_size = (size - 4) / 2; ?
                data_buff_size = size;
             }
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_UNICODE_STRING parameter '%s', decoded length %d > data buffer \
 size %d, will use the data buffer size instead.%c",
@@ -2439,7 +2439,7 @@ size %d, will use the data buffer size instead.%c",
          }
 
          // Sanity check, we should not get a negative element count.
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_UNICODE_STRING attribute '%s', decoded element count %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2450,7 +2450,7 @@ WARNING: For ENCODING_TYPE_UNICODE_STRING attribute '%s', decoded element count 
 
          // Handle the situation where more strings are in the input encoding
          // than what exist in the ref-attributes.
-         if ( (size_t)num_elements > num_items ) {
+         if ( (size_t)num_elements > num_items ) {    // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: Truncating array of ENCODING_TYPE_UNICODE_STRING from %d to %d elements for attribute '%s'!%c",
                      __LINE__, num_elements, num_items, FOM_name, THLA_NEWLINE );
@@ -2491,7 +2491,7 @@ WARNING: Truncating array of ENCODING_TYPE_UNICODE_STRING from %d to %d elements
 
             // Do a sanity check on the decoded length, it should not be negative.
             size_t length;
-            if ( decoded_count < 0 ) {
+            if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_UNICODE_STRING array element %d, attribute '%s', the decoded \
 length %d < 0, will use 0 instead.%c",
@@ -2503,7 +2503,7 @@ length %d < 0, will use 0 instead.%c",
 
             // Do a sanity check on the decoded length as compared to how much
             // data remains in the buffer.
-            if ( (size_t)length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_UNICODE_STRING array element %d, attribute '%s', the decoded \
 length %d > data buffer size %d, will use the data buffer size instead.%c",
@@ -2606,7 +2606,7 @@ length %d > data buffer size %d, will use the data buffer size instead.%c",
 
          // Do a sanity check on the decoded length, it should not be negative.
          size_t length;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_ASCII_STRING attribute '%s', decoded length %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2624,7 +2624,7 @@ WARNING: For ENCODING_TYPE_ASCII_STRING attribute '%s', decoded length %d < 0, w
             data_buff_size = 0;
          }
 
-         if ( length > data_buff_size ) {
+         if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_ASCII_STRING attribute '%s', decoded length %d > data buffer size \
 %d, will use the data buffer size instead.%c",
@@ -2694,7 +2694,7 @@ WARNING: For ENCODING_TYPE_ASCII_STRING attribute '%s', decoded length %d > data
          }
 
          // Sanity check, we should not get a negative element count.
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_ASCII_STRING attribute '%s', decoded element count %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2705,7 +2705,7 @@ WARNING: For ENCODING_TYPE_ASCII_STRING attribute '%s', decoded element count %d
 
          // Handle the situation where more strings are in the input encoding
          // than what exist in the ref-attributes.
-         if ( num_elements > num_items ) {
+         if ( num_elements > num_items ) {   // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: Truncating array of ENCODING_TYPE_ASCII_STRING from %d to %d elements for attribute '%s'!%c",
                      __LINE__, num_elements, num_items, FOM_name, THLA_NEWLINE );
@@ -2739,7 +2739,7 @@ WARNING: Truncating array of ENCODING_TYPE_ASCII_STRING from %d to %d elements f
 
             // Do a sanity check on the decoded length, it should not be negative.
             size_t length;
-            if ( decoded_count < 0 ) {
+            if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_ASCII_STRING array element %d, attribute '%s', the decoded \
 length %d < 0, will use 0 instead.%c",
@@ -2751,7 +2751,7 @@ length %d < 0, will use 0 instead.%c",
 
             // Do a sanity check on the decoded length as compared to how much
             // data remains in the buffer.
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_ASCII_STRING array element %d, attribute '%s', the decoded \
 length %d > data buffer size %d, will use the data buffer size instead.%c",
@@ -2854,7 +2854,7 @@ length %d > data buffer size %d, will use the data buffer size instead.%c",
 
          // Do a sanity check on the decoded length, it should not be negative.
          size_t length;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2871,7 +2871,7 @@ WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d < 0, wi
          } else {
             data_buff_size = 0;
          }
-         if ( length > data_buff_size ) {
+         if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d > data buffer size \
 %d, will use the data buffer size instead.%c",
@@ -2936,7 +2936,7 @@ WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded length %d > data 
          }
 
          // Sanity check, we should not get a negative element count.
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded element count %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2947,7 +2947,7 @@ WARNING: For ENCODING_TYPE_OPAQUE_DATA attribute '%s', decoded element count %d 
 
          // Handle the situation where more strings are in the input encoding
          // than what exist in the ref-attributes.
-         if ( num_elements > num_items ) {
+         if ( num_elements > num_items ) {   // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: Truncating array of ENCODING_TYPE_OPAQUE_DATA from %d to %d elements for attribute '%s'!%c",
                      __LINE__, num_elements, num_items, FOM_name, THLA_NEWLINE );
@@ -2981,7 +2981,7 @@ WARNING: Truncating array of ENCODING_TYPE_OPAQUE_DATA from %d to %d elements fo
 
             // Do a sanity check on the decoded length, it should not be negative.
             size_t length;
-            if ( decoded_count < 0 ) {
+            if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA array element %d, attribute '%s', the decoded \
 length %d < 0, will use 0 instead.%c",
@@ -2993,7 +2993,7 @@ length %d < 0, will use 0 instead.%c",
 
             // Do a sanity check on the decoded length as compared to how much
             // data remains in the buffer.
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Attribute::decode_string_from_buffer():%d \
 WARNING: For ENCODING_TYPE_OPAQUE_DATA array element %d, attribute '%s', the decoded \
 length %d > data buffer size %d, will use the data buffer size instead.%c",
