+++ source/TrickHLA/Parameter.cpp
@@ -1553,7 +1553,7 @@ void Parameter::decode_opaque_data_from_buffer()
 
       // Do a sanity check on the decoded length, it should not be negative.
       size_t length;
-      if ( decoded_count < 0 ) {
+      if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
          send_hs( stderr, "Parameter::decode_opaque_data_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA attribute '%s', decoded length %d < 0, will use 0 instead.%c",
                   __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -1571,7 +1571,7 @@ WARNING: For ENCODING_OPAQUE_DATA attribute '%s', decoded length %d < 0, will us
          data_buff_size = 0;
       }
 
-      if ( (size_t)length > data_buff_size ) {
+      if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
          send_hs( stderr, "Parameter::decode_opaque_data_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA attribute '%s', decoded length %d > data buffer \
 size %d, will use the data buffer size instead.%c",
@@ -2206,7 +2206,7 @@ void Parameter::decode_string_from_buffer()
 
          // Do a sanity check on the decoded length, it should not be negative.
          size_t length;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_UNICODE_STRING parameter '%s', decoded length %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2224,7 +2224,7 @@ WARNING: For ENCODING_UNICODE_STRING parameter '%s', decoded length %d < 0, will
             } else {
                data_buff_size = 0;
             }
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_UNICODE_STRING parameter '%s', decoded length %d > data buffer \
 size %d, will use the data buffer size instead.%c",
@@ -2302,7 +2302,7 @@ size %d, will use the data buffer size instead.%c",
 
          // Sanity check, we should not get a negative element count.
          size_t num_elements;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_UNICODE_STRING parmeter '%s', decoded element count %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2313,7 +2313,7 @@ WARNING: For ENCODING_UNICODE_STRING parmeter '%s', decoded element count %d < 0
 
          // Handle the situation where more strings are in the input encoding
          // than what exist in the ref-attributes.
-         if ( num_elements > num_items ) {
+         if ( num_elements > num_items ) {      // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: Truncating array of ENCODING_UNICODE_STRING from %d to %d elements for parameter '%s'!%c",
                      __LINE__, num_elements, num_items, FOM_name, THLA_NEWLINE );
@@ -2354,7 +2354,7 @@ WARNING: Truncating array of ENCODING_UNICODE_STRING from %d to %d elements for 
 
             // Do a sanity check on the decoded length, it should not be negative.
             size_t length;
-            if ( decoded_inner_count < 0 ) {
+            if ( decoded_inner_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_UNICODE_STRING array element %d of %d, parameter '%s', the decoded \
 length %d < 0, will use 0 instead.%c",
@@ -2366,7 +2366,7 @@ length %d < 0, will use 0 instead.%c",
 
             // Do a sanity check on the decoded length as compared to how much
             // data remains in the buffer.
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_UNICODE_STRING array element %d of %d, parameter '%s', the decoded \
 length %d > data buffer size %d, will use the data buffer size instead.%c",
@@ -2467,7 +2467,7 @@ length %d > data buffer size %d, will use the data buffer size instead.%c",
 
          // Do a sanity check on the decoded length, it should not be negative.
          size_t length;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_ASCII_STRING parmeter '%s', decoded length %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2485,7 +2485,7 @@ WARNING: For ENCODING_ASCII_STRING parmeter '%s', decoded length %d < 0, will us
             data_buff_size = 0;
          }
 
-         if ( length > data_buff_size ) {
+         if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_ASCII_STRING parameter '%s', decoded length %d > data buffer size \
 %d, will use the data buffer size instead.%c",
@@ -2556,7 +2556,7 @@ WARNING: For ENCODING_ASCII_STRING parameter '%s', decoded length %d > data buff
 
          // Sanity check, we should not get a negative element count.
          size_t num_elements;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_ASCII_STRING parameter '%s', decoded element count %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2567,7 +2567,7 @@ WARNING: For ENCODING_ASCII_STRING parameter '%s', decoded element count %d < 0,
 
          // Handle the situation where more strings are in the input encoding
          // than what exist in the ref-attributes.
-         if ( num_elements > num_items ) {
+         if ( num_elements > num_items ) {      // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: Truncating array of ENCODING_ASCII_STRING from %d to %d elements for parameter '%s'!%c",
                      __LINE__, num_elements, num_items, FOM_name, THLA_NEWLINE );
@@ -2602,7 +2602,7 @@ WARNING: Truncating array of ENCODING_ASCII_STRING from %d to %d elements for pa
 
             // Do a sanity check on the decoded length, it should not be negative.
             size_t length;
-            if ( decoded_inner_count < 0 ) {
+            if ( decoded_inner_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_ASCII_STRING array element %d, parameter '%s', the decoded \
 length %d < 0, will use 0 instead.%c",
@@ -2614,7 +2614,7 @@ length %d < 0, will use 0 instead.%c",
 
             // Do a sanity check on the decoded length as compared to how much
             // data remains in the buffer.
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_ASCII_STRING array element %d, parameter '%s', the decoded \
 length %d > data buffer size %d, will use the data buffer size instead.%c",
@@ -2717,7 +2717,7 @@ length %d > data buffer size %d, will use the data buffer size instead.%c",
 
          // Do a sanity check on the decoded length, it should not be negative.
          size_t length;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA parameter '%s', decoded length %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2734,7 +2734,7 @@ WARNING: For ENCODING_OPAQUE_DATA parameter '%s', decoded length %d < 0, will us
          } else {
             data_buff_size = 0;
          }
-         if ( length > data_buff_size ) {
+         if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA parameter '%s', decoded length %d > data buffer size \
 %d, will use the data buffer size instead.%c",
@@ -2798,7 +2798,7 @@ WARNING: For ENCODING_OPAQUE_DATA parameter '%s', decoded length %d > data buffe
 
          // Sanity check, we should not get a negative element count.
          size_t num_elements;
-         if ( decoded_count < 0 ) {
+         if ( decoded_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA parameter '%s', decoded element count %d < 0, will use 0 instead.%c",
                      __LINE__, FOM_name, decoded_count, THLA_NEWLINE );
@@ -2809,7 +2809,7 @@ WARNING: For ENCODING_OPAQUE_DATA parameter '%s', decoded element count %d < 0, 
 
          // Handle the situation where more strings are in the input encoding
          // than what exist in the ref-attributes.
-         if ( num_elements > num_items ) {
+         if ( num_elements > num_items ) {      // cppcheck-suppress [unsignedLessThanZero]
             send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: Truncating array of ENCODING_OPAQUE_DATA from %d to %d elements for parameter '%s'!%c",
                      __LINE__, num_elements, num_items, FOM_name, THLA_NEWLINE );
@@ -2844,7 +2844,7 @@ WARNING: Truncating array of ENCODING_OPAQUE_DATA from %d to %d elements for par
 
             // Do a sanity check on the decoded length, it should not be negative.
             size_t length;
-            if ( decoded_inner_count < 0 ) {
+            if ( decoded_inner_count < 0 ) {    // cppcheck-suppress [knownConditionTrueFalse]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA array element %d, parameter '%s', the decoded \
 length %d < 0, will use 0 instead.%c",
@@ -2856,7 +2856,7 @@ length %d < 0, will use 0 instead.%c",
 
             // Do a sanity check on the decoded length as compared to how much
             // data remains in the buffer.
-            if ( length > data_buff_size ) {
+            if ( length > data_buff_size ) {    // cppcheck-suppress [unsignedLessThanZero]
                send_hs( stderr, "Parameter::decode_string_from_buffer():%d \
 WARNING: For ENCODING_OPAQUE_DATA array element %d, parameter '%s', the decoded \
 length %d > data buffer size %d, will use the data buffer size instead.%c",
